refactor: split community page sub header from community base
