+++ app/web/src/features/communities/CommunityBase.tsx
@@ -1,23 +1,17 @@
-import { Breadcrumbs, Link as MuiLink, Typography } from "@material-ui/core";
-import { TabContext } from "@material-ui/lab";
 import Alert from "components/Alert";
 import CircularProgress from "components/CircularProgress";
-import TabBar from "components/TabBar";
-import JoinCommunityButton from "features/communities/CommunityPage/JoinCommunityButton";
 import {
-  COMMUNITY_TABS_A11Y_LABEL,
-  communityTabBarLabels,
   ERROR_LOADING_COMMUNITY,
   INVALID_COMMUNITY_ID,
 } from "features/communities/constants";
 import { useCommunity } from "features/communities/hooks";
 import { Community } from "proto/communities_pb";
-import { CommunityParent } from "proto/groups_pb";
 import React from "react";
-import { Link, useHistory, useParams } from "react-router-dom";
-import { CommunityTab, routeToCommunity } from "routes";
+import { useParams } from "react-router-dom";
+import { CommunityTab } from "routes";
 import makeStyles from "utils/makeStyles";
 
+import CommunityPageSubHeader from "./CommunityPage/CommunityPageSubHeader";
 import PageHeader from "./PageHeader";
 
 export const useCommunityBaseStyles = makeStyles((theme) => ({
@@ -31,16 +25,6 @@ export const useCommunityBaseStyles = makeStyles((theme) => ({
     display: "block",
     marginLeft: "auto",
     marginRight: "auto",
-  },
-  breadcrumbsContainer: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "space-between",
-  },
-  breadcrumbs: {
-    "& ol": {
-      justifyContent: "flex-start",
-    },
   },
 }));
 
@@ -71,15 +55,6 @@ export default function CommunityBase({
     data: community,
   } = useCommunity(communityId ?? +communityIdFromUrl);
 
-  const history = useHistory();
-
-  const { page = "overview" } = useParams<{ page: string }>();
-
-  // Use default tab from prop, otherwise derived state from community URL
-  const tab =
-    defaultTab ??
-    (page in communityTabBarLabels ? (page as CommunityTab) : "overview");
-
   if (!communityId && !communityIdFromUrl)
     return <Alert severity="error">{INVALID_COMMUNITY_ID}</Alert>;
 
@@ -96,54 +71,7 @@ export default function CommunityBase({
   return (
     <div className={classes.root}>
       {community.mainPage && <PageHeader page={community.mainPage} />}
-      <div className={classes.breadcrumbsContainer}>
-        <Breadcrumbs aria-label="breadcrumb" className={classes.breadcrumbs}>
-          {community.parentsList
-            .map((parent) => parent.community)
-            .filter(
-              (communityParent): communityParent is CommunityParent.AsObject =>
-                !!communityParent
-            )
-            .map((communityParent, index, array) =>
-              index === array.length - 1 ? (
-                <Typography
-                  variant="body1"
-                  color="textPrimary"
-                  key={`breadcrumb-${communityParent?.communityId}`}
-                >
-                  {communityParent.name}
-                </Typography>
-              ) : (
-                <MuiLink
-                  component={Link}
-                  to={routeToCommunity(
-                    communityParent.communityId,
-                    communityParent.slug
-                  )}
-                  key={`breadcrumb-${communityParent?.communityId}`}
-                >
-                  {communityParent.name}
-                </MuiLink>
-              )
-            )}
-        </Breadcrumbs>
-        <JoinCommunityButton community={community} />
-      </div>
-      <TabContext value={tab}>
-        <TabBar
-          ariaLabel={COMMUNITY_TABS_A11Y_LABEL}
-          setValue={(newTab) =>
-            history.push(
-              `${routeToCommunity(
-                community.communityId,
-                community.slug,
-                newTab === "overview" ? undefined : newTab
-              )}`
-            )
-          }
-          labels={communityTabBarLabels}
-        />
-      </TabContext>
+      <CommunityPageSubHeader community={community} defaultTab={defaultTab} />
       {children({ community, communitySlug })}
     </div>
   );
